Rename to `attributes_and_token_lists`
Remove the `ActionView` module nesting, and rename all files and
directories to reflect that un-nesting.
