--- lib/action_view/attributes_and_token_lists.rb
@@ -1,8 +1,0 @@
-require "action_view/attributes_and_token_lists/version"
-require "action_view/attributes_and_token_lists/engine"
-
-module ActionView
-  module AttributesAndTokenLists
-    # Your code goes here...
-  end
-end
