--- lib/action_view/attributes_and_token_lists/engine.rb
@@ -1,6 +1,0 @@
-module ActionView
-  module AttributesAndTokenLists
-    class Engine < ::Rails::Engine
-    end
-  end
-end
