--- lib/action_view/attributes_and_token_lists/version.rb
@@ -1,5 +1,0 @@
-module ActionView
-  module AttributesAndTokenLists
-    VERSION = "0.1.0"
-  end
-end
