+++ lib/attributes_and_token_lists.rb
@@ -1,0 +1,6 @@
+require "attributes_and_token_lists/version"
+require "attributes_and_token_lists/engine"
+
+module AttributesAndTokenLists
+  # Your code goes here...
+end
