+++ lib/attributes_and_token_lists/version.rb
@@ -1,0 +1,3 @@
+module AttributesAndTokenLists
+  VERSION = "0.1.0"
+end
