+++ lib/tasks/attributes_and_token_lists_tasks.rake
@@ -1,4 +1,4 @@
 # desc "Explaining what the task does"
-# task :action_view_attributes_and_token_lists do
+# task :attributes_and_token_lists do
 #   # Task goes here
 # end
