--- test/action_view/attributes_and_token_lists_test.rb
@@ -1,7 +1,0 @@
-require "test_helper"
-
-class ActionView::AttributesAndTokenListsTest < ActiveSupport::TestCase
-  test "it has a version number" do
-    assert ActionView::AttributesAndTokenLists::VERSION
-  end
-end
