+++ test/attributes_and_token_lists_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class AttributesAndTokenListsTest < ActiveSupport::TestCase
+  test "it has a version number" do
+    assert AttributesAndTokenLists::VERSION
+  end
+end
