+++ test/dummy/config/application.rb
@@ -7,7 +7,7 @@ require "rails/test_unit/railtie"
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
-require "action_view/attributes_and_token_lists"
+require "attributes_and_token_lists"
 
 module Dummy
   class Application < Rails::Application
